fix: move folder 'schemas' folder is now inside 'src'
