+++ src/removeCoioteFormat.ts
@@ -12,7 +12,7 @@ import type {
 	Temperature_3303,
 	Temperature_3303_urn,
 } from '@nordicsemiconductor/lwm2m-types'
-import { type Config_50009, Config_50009_urn } from '../schemas/Config_50009.js'
+import { type Config_50009, Config_50009_urn } from './schemas/Config_50009.js'
 
 import { type assetTrackerObjects } from './getAssetTrackerV2Objects.js'
 import { setCustomFormat } from './utils/setCustomFormat.js'
